requirements.txt
flask
pyGeckoWiiU